Refactor: Update app launch screen and theme colors
This commit updates the app's launch screen and theme colors.

The launch screen background color is changed to a custom `baseColor` (#EFE8D7). The app icon is now displayed on the launch screen. The `launch_background.xml` for API v21 is removed as it's no longer necessary.

The `baseColor` is also applied as the `seedColor` for both light and dark themes in the `MaterialApp` widget.

New adaptive launcher icons and their corresponding background and foreground image assets are added. The existing launcher icons are replaced with the new ones.