Refactor: Relocate remove buttons in checklist and task items
The remove button in both checklist and task item widgets has been moved from the trailing position to the leading position.

In the checklist item, the trailing widget is now a chevron right icon.
In the task item, the checkbox has been moved to the trailing position.
Padding in the checklist item has also been adjusted.